préférence dans la page de stats

+++ app/src/main/java/com/example/thirstyquest/data/CategoryCollection.kt
@@ -4,5 +4,5 @@ data class Category(
     val name: String,
     val level: Int = 0,
     val points: Double = 0.0,
-    val total: Double = 0.0,
+    val total: Int = 0,
 )
